Enable playing the audio at correct times
